+++ scripts/get_model.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+
+curl -iv 'http://localhost:9444/models/default/samples?start=0&limit=100'