+++ scripts/train_model.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+
+curl -iv -X POST 'http://localhost:9444/models/default/train'